requirements.txt
streamlit
wordcloud
matplotlib
nltk
openai
PyPDF2
python-dotenv